Move signal_scheduler::{Profile,Sample} to top-level
