+++ ext/pf2/src/lib.rs
@@ -4,7 +4,9 @@ extern crate serde_derive;
 
 mod ruby_init;
 
+mod profile;
 mod profile_serializer;
+mod sample;
 mod sample_collector;
 mod signal_scheduler;
 mod util;
